+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - [322](https://github.com/vegaprotocol/vegacapsule/pull/322) Improve stop command & update vega core dependency
 - [329](https://github.com/vegaprotocol/vegacapsule/pull/329) Add template functions for generating `IPFS peer ids`.
 - [337](https://github.com/vegaprotocol/vegacapsule/issues/337) Allow > 9 and < 100 nodes to be created.
+- [348](https://github.com/vegaprotocol/vegacapsule/issues/348) Do not parse wallet initialisation output anymore
 
 ### 🐛 Fixes
 - [167](https://github.com/vegaprotocol/vegacapsule/issues/167) Fix validators filter in tendermint generator
@@ -91,4 +92,4 @@
 - [117](https://github.com/vegaprotocol/vegacapsule/pull/117) - Fix nil dereference panics in config
 - [41](https://github.com/vegaprotocol/vegacapsule/issues/40) - Persist the network state after it's generated in bootstrap command
 - [86](https://github.com/vegaprotocol/vegacapsule/issues/86) - Allow overriding config options that default true with falue
-- [323](https://github.com/vegaprotocol/vegacapsule/pull/323) - 
+- [323](https://github.com/vegaprotocol/vegacapsule/pull/323) -